Separate unit letters related stuff from FlapFunctions.cpp
