+++ letters.h
@@ -1,0 +1,11 @@
+#ifndef LETTERS_H
+#define LETTERS_H
+
+#include "stringHandling.h"
+
+
+int getSuggestedOffset(int letterIndex);
+int translateLetterToIndex(char letterchar);
+char translateIndextoLetter(int index);
+
+#endif // LETTERS_H
